pyproject.toml
[tool.poetry]
name = "instagram mini"
version = "0.1.0"
description = "assesment project"
authors = ["sakthiRatnam <sakthiratnam050@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "3.0.5"
django-allauth = "0.44.0"
django-cors-headers = "3.7.0"
django-cors-middleware = "1.5.0"
django-filter = "2.4.0"
django-rest-auth = "0.9.5"
django-rest-swagger = "2.2.0"
django-robots = "4.0"
django-url-filter = "0.3.15"
djangorestframework = "3.12.4"
psycopg2 = "2.8.6"
uvicorn = "0.14.0"
PyJWT = "2.1.0"
gunicorn = "20.1.0"
razorpay = "^1.2.0"
qrcode = "^7.3"
reportlab = "^3.6.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
